% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aluguel2025sjdr}
\alias{aluguel2025sjdr}
\title{Rental property data of Sao Joao del Rei city in Brazil}
\format{
A dataframe with 191 observations on 11 variables.
\describe{
\item{bairro}{neighborhood of the property}
\item{numero_quartos}{number of bedrooms}
\item{numero_banheiros}{number of bathrooms}
\item{vagas_carros}{number of car parking spaces}
\item{area_gourmet}{if the property has gourmet area}
\item{mobiliado}{if the property is furnished}
\item{varanda}{if the property has balcony}
\item{imobiliaria}{if the property is advertised in real estate}
\item{tipo}{type of property, including house, apartment, and others}
\item{preco}{monthly rental price}
\item{Link}{website of the advertising}
}
}
\source{
Data collected by Lucas Emanuel Pereira de Melo, Rafael
Rodrigues da Costa Carvalho, and Diogo Oliveira Murari Rabelo
}
\usage{
aluguel2025sjdr
}
\description{
Dataset containing information on property rentals in Sao Joao del Rei
for the year 2025, including prices, property features, and location.
}
\examples{
data(aluguel2025sjdr)
head(aluguel2025sjdr)
summary(aluguel2025sjdr$preco)
}
\keyword{datasets}
