% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apartamentos2024mg}
\alias{apartamentos2024mg}
\title{Apartment prices in Minas Gerais in 2024}
\format{
A dataframe with 632 observations on 14 variables.
\describe{
\item{Cidade}{apartment city}
\item{Bairro}{apartment neighborhood}
\item{Area}{apartment area in m^2}
\item{Valor}{apartment price}
\item{Quartos}{number of bedrooms}
\item{Banheiros}{number of bathrooms}
\item{Vaga}{number of car parking spaces}
\item{Varanda}{if the property has balconi}
\item{Suite}{if the property has suite}
\item{Area.Gourmet}{if the property has gourmet area}
\item{Terraco}{if the property has terrace}
\item{Sala}{number of livin grooms}
\item{Copa}{if the property has dining room}
\item{Piscina}{if the property has swimming pool}
}
}
\source{
Data collected by Andre Vinicius Amaral Leal, Andressa Marileia Ladeira,
Caio Castro Maia, Eliza Cabrera da Silva, Julia Vitoria Martimiano,
Ketlly Lopes Leite, and Matheus Oliveira Cardoso
}
\usage{
apartamentos2024mg
}
\description{
Dataset containing information about apartments available for sale in Minas Gerais,
for the year 2024, including sale prices, property characteristics, and location.
}
\examples{
data(apartamentos2024mg)
head(apartamentos2024mg)
plot(apartamentos2024mg$Area, apartamentos2024mg$Valor)
}
\keyword{datasets}
