% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carrosusados2025webmotors}
\alias{carrosusados2025webmotors}
\title{Resale prices of used cars - Webmotors}
\format{
A dataframe with 200 observations on 9 variables.
\describe{
\item{Marca}{car manufacturer}
\item{Carro}{car model}
\item{Ano}{year of the car}
\item{Km}{mileage}
\item{Cambio}{automatic or manual transmission}
\item{Motor}{motor displacement}
\item{Valvulas}{number of engine valves}
\item{Combustivel}{type of comustivel, flex means hibrid of gasoline and ethanol}
\item{Preco}{price}
}
}
\source{
Data collected by Bryan Matheus Silva, Lua Teixeira Guimaraes Alves Carneiro,
and Vitoria Aparecida Morais
}
\usage{
carrosusados2025webmotors
}
\description{
Dataset containing information about used cars collected from the Webmotors platform
in 2025, including model, manufacturer, prices, vehicle characteristics,
year, and mileage.
}
\examples{
data(carrosusados2025webmotors)
head(carrosusados2025webmotors)
table(carrosusados2025webmotors$Marca)
}
\keyword{datasets}
