% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celularesnovos2025}
\alias{celularesnovos2025}
\title{New cellphones prices}
\format{
A dataframe with 200 observations on 10 variables.
\describe{
\item{modelo}{cellphone model}
\item{marca}{cellphone manufacturer}
\item{ram}{cellphone ram memory}
\item{armazenamento}{storage capacity in Gb}
\item{camerah}{camera resolution height in pixels}
\item{cameral}{camera resolution width in pixels}
\item{ano}{release year}
\item{resolucao}{screen resolution}
\item{bateria}{batery capacity in mAh}
\item{Preco}{cellphone price}
}
}
\source{
Data collected by Alan Sebastiao do Nascimento, Ana Julia Patricio Melo,
Fernanda Dias do Valle, and Thiago Vinicius Portes
}
\usage{
celularesnovos2025
}
\description{
Dataset containing information about new cell phones available in 2025,
including price, model, and technical specifications.
}
\examples{
data(celularesnovos2025)
head(celularesnovos2025)
boxplot(celularesnovos2025$Preco ~ celularesnovos2025$marca)
}
\keyword{datasets}
