% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celularesusados}
\alias{celularesusados}
\title{Resale price of used cell phones}
\format{
A dataframe with 200 observations on 8 variables.
\describe{
\item{modelo}{cellphone model}
\item{marca}{cellphone manufacturer}
\item{anolancamento}{release year}
\item{armazenamento}{storage capacity in Gb}
\item{estado}{condition/state of preservation}
\item{nota_fiscal}{if the cell phone has a receipt}
\item{fonte}{reseler platform}
\item{preco}{resale price}
}
}
\source{
Data collected by Bruna Machado Melo, Gabriele Campos Pedrosa,
Luiza Ferreira Carregal, and Sara Zelia Luiz Lemos
}
\usage{
celularesusados
}
\description{
Dataset containing information about used cell phones, including price,
technical specifications, condition, and device features.
}
\examples{
data(celularesusados)
head(celularesusados)
hist(celularesusados$preco)
}
\keyword{datasets}
