% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{computadores}
\alias{computadores}
\title{Laptop computers}
\format{
A data frame with 312 observations on 6 variables.
\describe{
\item{marca}{brand name}
\item{processador}{processor model}
\item{ram}{RAM memory in GB}
\item{armazenamento}{storage capacity in GB}
\item{tipo_armazenamento}{storage type}
\item{preco}{price in BRL}
}
}
\source{
Data collected by Lucas Teixeira Faccion and Pedro Santos Oliveira
}
\usage{
computadores
}
\description{
Dataset containing information about laptop computers available in the
Brazilian market, including brand, processor, memory, storage, and price.
}
\examples{
data(computadores)
head(computadores)
hist(computadores$preco)
}
\keyword{datasets}
