% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feijoes}
\alias{feijoes}
\title{Feijoes (beans)}
\format{
A dataframe with 250 observations on 5 variables.
\describe{
\item{feijao}{bean specie/variety}
\item{comprimento}{bean length in mm}
\item{largura}{bean width in mm}
\item{espessura}{bean thickness in mm}
\item{massa}{mass in g}
}
}
\source{
Data collected by Breno Ferreira Barrados Chorro, Helio Tavares de Sa Neto,
Carlos Henrique De Magalhaes, Raul Fernando De Castro Junior, Larissa Faleiro Malta,
Samyres Damasceno Margotti, and Joao Vitor de Siqueira Roque
}
\usage{
feijoes
}
\description{
Dataset with measurements on distinct beans found in brazilian market
}
\examples{
data(feijoes)
head(feijoes)
table(feijoes$feijao)
}
\keyword{datasets}
