% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{folhasfrutas}
\alias{folhasfrutas}
\title{Fruit Tree Leaves}
\format{
A data frame with 235 observations on 9 variables.
\describe{
\item{especie}{tree specie}
\item{area}{leave area}
\item{perimeter}{leave perimeter}
\item{radius.mean}{average leave radius}
\item{radius.sd}{standard deviation of leave radius}
\item{radius.min}{minimum leave radius}
\item{radius.max}{maximum leave radius}
\item{majoraxis}{major axis leave dimension}
\item{eccentricity}{leave eccentricity}
}
}
\source{
Data collected by Andre Vinicius Amaral Leal, Andressa Marileia Ladeira,
Caio Castro Maia, Eliza Cabrera da Silva, Julia Vitoria Martimiano,
Ketlly Lopes Leite, Matheus Oliveira Cardoso, and Robson Bruno Dutra Pereira
}
\usage{
folhasfrutas
}
\description{
Dataset on measurements obtained through image analysis of fruit tree leaves
}
\examples{
data(folhasfrutas)
head(folhasfrutas)
}
\keyword{datasets}
