% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soja}
\alias{soja}
\title{Soybean seed morphology}
\format{
A data frame with 540 observations on 32 variables.
\describe{
\item{area}{seed area}
\item{area_ch}{convex hull area}
\item{perimeter}{seed perimeter}
\item{radius_mean}{mean radius}
\item{radius_min}{minimum radius}
\item{radius_max}{maximum radius}
\item{radius_sd}{standard deviation of radius}
\item{diam_mean}{mean diameter}
\item{diam_min}{minimum diameter}
\item{diam_max}{maximum diameter}
\item{major_axis}{major axis length}
\item{minor_axis}{minor axis length}
\item{caliper}{caliper diameter}
\item{length}{seed length}
\item{width}{seed width}
\item{radius_ratio}{ratio between maximum and minimum radius}
\item{theta}{orientation angle}
\item{eccentricity}{seed eccentricity}
\item{form_factor}{form factor}
\item{narrow_factor}{narrow factor}
\item{asp_ratio}{aspect ratio}
\item{rectangularity}{rectangularity}
\item{pd_ratio}{perimeter-diameter ratio}
\item{plw_ratio}{perimeter-length-width ratio}
\item{solidity}{solidity}
\item{convexity}{convexity}
\item{elongation}{elongation}
\item{circularity}{circularity}
\item{circularity_haralick}{Haralick circularity}
\item{circularity_norm}{normalized circularity}
\item{coverage}{seed coverage}
\item{cultura}{soybean cultivar (response variable for classification)}
}
}
\source{
Data collected by Bionda Letícia Da Costa Carvalho,
João Pedro Terrim Leite, Rafaela Rayane Rodrigues Santos,
and Sandy Aparecida Ribeiro
}
\usage{
soja
}
\description{
Dataset containing morphological measurements obtained through image analysis
of soybean seeds from different cultivars, for classification purposes.
}
\examples{
data(soja)
head(soja)
table(soja$cultura)
}
\keyword{datasets}
