% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperos}
\alias{temperos}
\title{Spice and herb leaf morphology}
\format{
A data frame with 300 observations on 13 variables.
\describe{
\item{especie}{plant species (response variable for classification):
Espinafre, Manjericao, Melissa, Oregano, or Salvia}
\item{area}{leaf area in pixels}
\item{perimeter}{leaf perimeter in pixels}
\item{length}{leaf length in pixels}
\item{width}{leaf width in pixels}
\item{major_axis}{major axis length}
\item{minor_axis}{minor axis length}
\item{eccentricity}{leaf eccentricity}
\item{asp_ratio}{aspect ratio}
\item{elongation}{elongation}
\item{circularity}{circularity}
\item{solidity}{solidity}
\item{form_factor}{form factor}
}
}
\source{
Data collected by Isabel Morena De Oliveira Silva,
Marcos Felipe Resgalla Da Matta Arvelos,
Mariana Christofaro Brighenti,
and Meirieli Aparecida Oliveira Sousa
}
\usage{
temperos
}
\description{
Dataset containing morphological measurements obtained through image analysis
of spice and herb leaves from different species, for classification purposes.
}
\examples{
data(temperos)
head(temperos)
table(temperos$especie)
}
\keyword{datasets}
